purpurite/requirements.txt
bernhard>=0.1
celery>=3.0.24
pip==1.4.1
protobuf>=2.5.0
pysmb>=1.1.7
redis>=2.8.0
sqlalchemy>=0.8.2
git+ssh://git@github.com/Deca-Technologies/painite.git#egg=painite-0.2.2
